pyproject.toml
[project]
name = "azure-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "numpy>=2.3.1",
    "openai>=1.90.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "python-dotenv>=1.1.0",
]
